Right Arrow convenience functionality
The right arrow key can be used to populate the input box /text area with the currently selected text. This is handy for autocomplete type scenarios where you have long but similar fields that you may want to skip to. An example of this is when you have table.navigationfield.field1 and would like to skip to the end and select table.navigationfield.field9. With the right arrow feature just get to the first entry (table.navigationfield.field1) and then type right arrow + backspace + 9 and you don't have to scroll all the way through